chore(discover) Move TableDataRow type to utils
This results in fewer cross view imports, and co-locates the result type
with the component that generally creates the results.
