+++ src/sentry/static/sentry/app/components/events/interfaces/spans/spanDetail.tsx
@@ -6,7 +6,6 @@ import * as Sentry from '@sentry/react';
 import {Organization, SentryTransactionEvent} from 'app/types';
 import {Client} from 'app/api';
 import {IconWarning} from 'app/icons';
-import {TableDataRow} from 'app/views/eventsV2/table/types';
 import {assert} from 'app/types/utils';
 import {generateEventSlug, eventDetailsRoute} from 'app/utils/discover/urls';
 import {getParams} from 'app/components/organizations/globalSelectionHeader/getParams';
@@ -19,8 +18,9 @@ import Link from 'app/components/links/link';
 import LoadingIndicator from 'app/components/loadingIndicator';
 import Pill from 'app/components/pill';
 import Pills from 'app/components/pills';
-import getDynamicText from 'app/utils/getDynamicText';
 import space from 'app/styles/space';
+import getDynamicText from 'app/utils/getDynamicText';
+import {TableDataRow} from 'app/utils/discover/discoverQuery';
 import withApi from 'app/utils/withApi';
 
 import {ProcessedSpanType, RawSpanType, ParsedTraceType, rawSpanKeys} from './types';
